pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dwu"
version = "0.3.2"
description = "A dynamic wallpaper updater"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "starrieste"}
]

dependencies = [
    "click",
    "httpx",
    "selectolax",
    "pillow"
]

[project.scripts]
dwu = "dwu.__main__:main"

[project.urls]
Repository = "https://github.com/starrieste/dwu"

[tool.setuptools]
packages = ["dwu"]
